add change state  method to controller
