+++ epamticket/src/com/epam/ticket/data/EpamTicketStateHolder.java
@@ -1,0 +1,25 @@
+package com.epam.ticket.data;
+
+import java.io.Serializable;
+
+public class EpamTicketStateHolder implements Serializable {
+    private String newState;
+    private String comment;
+
+    public String getNewState() {
+        return newState;
+    }
+
+    public void setNewState(String newState) {
+        this.newState = newState;
+    }
+
+    public String getComment() {
+        return comment;
+    }
+
+    public void setComment(String comment) {
+        this.comment = comment;
+    }
+
+}
